+++ app/src/main/java/com/capstone/nongglenonggle/app/Screens.kt
@@ -4,7 +4,7 @@ package com.capstone.nongglenonggle.app
 //어디로 이동하는지 명확하지 않은 경우가 많음
 //이를 방지하기 위해 컴파일 타임에 유효한 경로를 사용할 수 있도록 모든 경로 타입을 정의하여 구조화함
 sealed class Screens(val route: String) {
-    object StartNonggleScreen: Screens("StartNonggle")
+    object Login: Screens("Login")
     object Signup: Screens("SignUp") {
         object SetType: Screens("SignUp/SetType")
         object Step2: Screens("SignUp/Step2")
