+++ app/src/main/java/com/capstone/nongglenonggle/presentation/view/login/LoginActivity.kt
@@ -2,18 +2,10 @@ package com.capstone.nongglenonggle.presentation.view.login
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
-import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.compose.setContent
-import androidx.activity.result.IntentSenderRequest
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.lifecycleScope
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.rememberNavController
-import com.google.android.gms.auth.api.identity.Identity
+import com.capstone.nongglenonggle.app.NonggleNavHost
 import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.launch
+import com.capstone.nongglenonggle.core.design_system.NongleTheme
 
 @AndroidEntryPoint
 class LoginActivity : ComponentActivity() {
@@ -22,31 +14,8 @@ class LoginActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
 
         setContent {
-            val navController = rememberNavController()
-            NavHost(navController = navController, startDestination = "login") {
-                composable("login") {
-                    val viewModel = hiltViewModel<LoginViewModel>()
-
-                    val launcher = rememberLauncherForActivityResult(
-                        contract = ActivityResultContracts.StartIntentSenderForResult(),
-                        onResult = { result ->
-                            if (result.resultCode == RESULT_OK) {
-                                result.data?.let {
-                                    viewModel.setEvent(LoginContract.Event.OnGoogleSignInResult(it))
-                                }
-                            }
-                        }
-                    )
-
-                    LoginScreen(
-                        viewModel = viewModel,
-                        onLaunchGoogleSignIn = { intentSender ->
-                            launcher.launch(
-                                IntentSenderRequest.Builder(intentSender).build()
-                            )
-                        }
-                    )
-                }
+            NongleTheme {
+                NonggleNavHost()
             }
         }
